+++ src/main/java/interface_adapter/search/SearchState.java
@@ -3,7 +3,7 @@ package interface_adapter.search;
 import java.util.ArrayList;
 import java.util.List;
 
-import entity.Listing;
+import entity.listing.Listing;
 
 /**
  * The State information representing book characterization for search.
